requirements.txt
pyats
genie
unicon
pyyaml
jinja2
allure-behave==2.15.3
allure-pytest==2.15.3
allure-python-commons==2.15.3
astroid==4.0.2
async-generator==1.10
attrs==25.4.0
beautifulsoup4==4.14.3
behave==1.3.3
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.6
cucumber-expressions==19.0.0
cucumber-tag-expressions==9.1.0
dill==0.4.1
distlib==0.4.0
filelock==3.25.2
h11==0.16.0
identify==2.6.16
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
isort==8.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jwt==1.4.0
mailosaur==8.1.0
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.10.0
outcome==1.3.0.post0
packaging==26.0
parse==1.21.1
parse_type==0.6.6
pillow==12.1.1
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
pycparser==3.0
Pygments==2.20.0
PyJWT==2.12.0
pylint==4.0.5
pylint-pytest==1.1.8
PySocks==1.7.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-discovery==1.2.1
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.0
rpds-py==0.30.0
selenium==4.41.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
timer==0.3.0
tomlkit==0.14.0
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
virtualenv==21.2.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
zipp==3.23.0